APIs/appwrite.io/server/0.9.3/python-aiohttp-client/openapi_server/controllers/health_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def health_get(request: web.Request, ) -> web.Response:
    """Get HTTP

    Check the Appwrite HTTP server is up and responsive.


    """
    return web.Response(status=200)


async def health_get_anti_virus(request: web.Request, ) -> web.Response:
    """Get Anti virus

    Check the Appwrite Anti Virus server is up and connection is successful.


    """
    return web.Response(status=200)


async def health_get_cache(request: web.Request, ) -> web.Response:
    """Get Cache

    Check the Appwrite in-memory cache server is up and connection is successful.


    """
    return web.Response(status=200)


async def health_get_db(request: web.Request, ) -> web.Response:
    """Get DB

    Check the Appwrite database server is up and connection is successful.


    """
    return web.Response(status=200)


async def health_get_queue_certificates(request: web.Request, ) -> web.Response:
    """Get Certificate Queue

    Get the number of certificates that are waiting to be issued against [Letsencrypt](https://letsencrypt.org/) in the Appwrite internal queue server.


    """
    return web.Response(status=200)


async def health_get_queue_functions(request: web.Request, ) -> web.Response:
    """Get Functions Queue

    


    """
    return web.Response(status=200)


async def health_get_queue_logs(request: web.Request, ) -> web.Response:
    """Get Logs Queue

    Get the number of logs that are waiting to be processed in the Appwrite internal queue server.


    """
    return web.Response(status=200)


async def health_get_queue_tasks(request: web.Request, ) -> web.Response:
    """Get Tasks Queue

    Get the number of tasks that are waiting to be processed in the Appwrite internal queue server.


    """
    return web.Response(status=200)


async def health_get_queue_usage(request: web.Request, ) -> web.Response:
    """Get Usage Queue

    Get the number of usage stats that are waiting to be processed in the Appwrite internal queue server.


    """
    return web.Response(status=200)


async def health_get_queue_webhooks(request: web.Request, ) -> web.Response:
    """Get Webhooks Queue

    Get the number of webhooks that are waiting to be processed in the Appwrite internal queue server.


    """
    return web.Response(status=200)


async def health_get_storage_local(request: web.Request, ) -> web.Response:
    """Get Local Storage

    Check the Appwrite local storage device is up and connection is successful.


    """
    return web.Response(status=200)


async def health_get_time(request: web.Request, ) -> web.Response:
    """Get Time

    Check the Appwrite server time is synced with Google remote NTP server. We use this technology to smoothly handle leap seconds with no disruptive events. The [Network Time Protocol](https://en.wikipedia.org/wiki/Network_Time_Protocol) (NTP) is used by hundreds of millions of computers and devices to synchronize their clocks over the Internet. If your computer sets its own clock, it likely uses NTP.


    """
    return web.Response(status=200)
